Temporary workaround for stack test
Formerly "stack test" worked, where the cabal file stanza
for the test suite omitted the package dependency for this
library, and defined the TEST variable that conditionally
exports a function only used for testing.

In the new branch, the mysterious mechanism seems broken.
For now, always export the function, and add the package
dependency.